+++ packages/xchain-thorchain-amm/src/types.ts
@@ -14,14 +14,12 @@ export type ExecuteSwap = {
   destinationAddress?: Address
   memo: string
   feeOption?: FeeOption
-  waitTimeSeconds: number
   walletIndex: number
 }
 
 export type TxSubmitted = {
   hash: string
   url: string
-  waitTimeSeconds: number
 }
 
 export type LiquidityPosition = {